+++ CHANGELOG.md
@@ -11,6 +11,7 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 - Do not start print progress bar before checking whether output file exists
 - Only derive the output file name during decompression when the input file has a `zst` extension
+- Refuse reading from stdin, if it is a terminal. Use `--force` to disable the check
 
 ## [0.4.4-cli]